Refactoring #1 Phase 1 - Part 1
Creation of a Cache system, mainly to store the biggest lists to decrease the transferts between the ASP.NET Client and the WCF Server. First, the main lists are stored - Projects,  Users, Teams, Departments. The current timeout is set to 10:00 minutes. The edition (add, update, delete) is not implemented yet.

The interactions with the sessions are stored in a generic controller to reduce the amount of useless code. In the part 2, other improvements regarding the dataview and others will be done, as well as a stronger place of the cache.