Refactor #2 Phase 1 - Part 2
Added more consistency to the cache system and transferred a major part of the actions to the SessionWrapper (containing all required operations). The third part of this refactor will pass by the cleaning of the SessionWrapper and Cache functions. Also, this time, cache for tasks have been added, meaning it will be split between project tasks and user tasks.