Client improvements - Browser
Created partial views to upload files from the client to the ASP.NET Server, in the selected repository. Added folders visualization and management (begin of ) plus travels between the documents.
Provided a way to upload files to the current repository, with a (disabled) view to confirm the success of the operation. This idea will be expanded in a far future.

Improved profile, team and project avatar / icon upload. Nicer UI in general.

**** Note: The next commits will touch the mailing system by adding a groupped-mailing (especially if you click on the "send mail to" from the department view.) In addition, a table will be created for the element Filr in which all data regarding ownership, updates and so on will be registered. But first, the message and notification system need a little rework (especially the UI) by including sent mails, a separation between the items.

A lots of Scripts are stored in the Layout, these will be modified during the next Refactoring session (should coming in one or two weeks at least). The buttons-like glyphicons are kept to identity actions and controller-related hyperlinks.

Also, the UWP should be started once the Browser refactoring session is either started or ended. It will be decided in the future.