Client improvements - Browser
UI Improvements regarding the Messages and Notifications. Added features to the department like users in this department - with their teams and so on. New element is groupped mailing, from which a user can send to several users or even departments a message through mailbox.

Added multi-team supports on the Project management, several teams can be on one project - especially Testers / Developpers and so on - and new notifications are sent if the administrator has changed.

Now starting a big site, file tracking and comments. At the moment, only a basis in XML is offered yet non functional. New tables regarding the files are on the way, yet will be in use once the XML provider is online.

**** Note :  The chat system is likely to be changed based on this XML providing elements rather than database saving directly. So, regarding the effectiveness and the conception, the solution will be taken at the end of the File Tracking System.

In nearly one week or two, the administrative side of the project will be reupdated for various reasons. Models not up-to-date on the other side, UI revamp and so on.