Server Side - Improvements Started integrating the concept of Dashboard inside the Workstation Administration Browser. ATM, it only shows the changes per projects, over all the files, yet in the next patches it will display them inside divs with some other details and so on.
On the client side, abandonning ReactJS for TypeScript and added somes features like total changes on this file.

At the same time, started a new brand part, oriented towards Go this time, for real-time XML Reading for the comment sections, instead of using ASP.NET. Of course, it is still under construction so it might not be effective right now.

**** Note: The next patches will mainly focus on the Go / Admin interactions, but first it will focus on the Go part simply. After that, a new SignalR Implement might be done to do Real-Time checking on the database from the dashboarding systems.