Server Side - Improvements Upgrade several tables like ChangeSet to give, right now, correct data such as ChangeSet commit date, latest committeer. At the same time, the dashboard has been improved on several points: • First a bug fix on the client side that generated the wrong paths. It has been corrected • Improving the Comment Drone, now the latter finds the comment files recursively and so sends everything back to the dashboard
On the server side, all the comment files are received and the changesets are too. Right now, the changesets are displayed only if the comment activities exist, yet, in the future, a cross-check to verify if it exists on one side or the other or even both should be implemented in the controller.

**** Note: The drone uses absolute paths right now, but a workaround should be found in the future. As for the way the dashboard is displayed, huge improvements might appear in the next commits. This will regroup all files by folders and, so, lower the final charge.
Now that the commit system is more or less functional, ability to set a description, a parent commit etc., shall be one of the next big thing to implement.